+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/shayanriyaz/anaconda3/envs/BestGeese/bin/python"
+}
