+++ backend/app.py
@@ -1,8 +1,7 @@
-from flask import Flask,Response,request,jsonify
+from flask import Flask
 from http import HTTPStatus
 from database.db import initialize_db
-from database.models import Question
-import json
+from resources.question import questions
 
 
 app = Flask(__name__)
@@ -26,49 +25,11 @@ app.config['MONGODB_SETTINGS'] = {
 }
 
 initialize_db(app)
-
+app.register_blueprint(questions)
 
 @app.route('/')
 def hello():
     return {'Hello':'World'}
     
-@app.route('/questions')
-def get_questions():
-    questions = Question.objects().to_json()
-    return Response(questions,
-                    mimetype='application/json',
-                    status=200)
-
-
-@app.route('/questions',methods=['POST'])
-def add_question():
-    body = request.get_json()
-    question = Question(**body).save()
-    id = question.id
-    return{'id':str(id)},HTTPStatus.OK
-
-
-@app.route('/questions/<id>',methods=['PUT'])
-def update_question(id):
-    body = request.get_json()
-    
-    if body is None:
-        return {'Error':'Question not found'},HTTPStatus.NO_CONTENT
-    
-    Question.get(id=id).update(**body)
-    return '',HTTPStatus.OK
-
-@app.route('/questions/<id>',methods=['DELETE'])
-def delete_question(id):
-    question = Question.objects.get(id=id).delete()
-    return '',HTTPStatus.OK
-
-@app.route('/questions/<id>')
-def get_question(id):
-    questions = Question.objects.get(id=id).to_json()
-    return Response(questions,
-                    mimetype="application/json",
-                    status = 200)
-
 if __name__ == "__main__":
     app.run(debug=True)
